Compass and Wind setup in a separate module.
