+++ main/setup/SubMenuCompassWind.h
@@ -1,0 +1,6 @@
+
+#pragma once
+
+class SetupMenu;
+
+void options_menu_create_compasswind(SetupMenu *top);
